// src/client/enums/GameEvent.ts
export enum GAME_EVENT {
  startGame = "startGame",
  moveCard = "moveCard",
  changeP2Health = "changeP2Health",
  changeP1Health = "changeP1Health",
  changeP2AP = "changeP2AP",
  changeP1AP = "changeP1AP",
  resetGame = "resetGame",
  focusCard = "focusCard",
  clearCardInFocus = "clearCardInFocus",
  selectCard = "selectCard",
  clearSelectedCard = "clearSelectedCard",
  setPileInView = "setPileInView",
  clearPileInView = "clearPileInView",
  shuffleTargetPile = "shuffleTargetPile",
  flipCard = "flipCard",
  closeSetDecksModal = "closeSetDecksModal",
  openSetDecksModal = "openSetDecksModal",
  increaseCardAttackModifier = "increaseCardAttackModifier",
  decreaseCardAttackModifier = "decreaseCardAttackModifier",
  increaseCardOtherModifier = "increaseCardOtherModifier",
  decreaseCardOtherModifier = "decreaseCardOtherModifier",
  increaseCardCooldown = "increaseCardCooldown",
  decreaseCardCooldown = "decreaseCardCooldown",
  rollDie = "rollDie",
  sendChatMessage = "sendChatMessage",
  switchSide = "switchSide",
  setP1Viewing = "setP1Viewing",
  setP2Viewing = "setP2Viewing",
  plunder = "plunder",
  //
  nextPhase = "nextPhase",
  setRpsChoice = "setRpsChoice",
  mulligan = "mulligan",
  //
  playerInput = "playerInput",
  conscript = "conscript",
}